Extract Cuts module from Packet class
